Shares PrefundAccounts method across tests.
